+++ app/src/main/java/com/example/iweathercompose/data/dto/ForecastWeather.kt
@@ -34,7 +34,7 @@ data class ForecastWeather(
     ) {
         data class Coord(
             @SerializedName("lat")
-            val lat: Int? = null,
+            val lat: Double? = null,
             @SerializedName("lon")
             val lon: Double? = null
         )
